# Use atomics instead of synchronized for TimeWindowQuantiles