--- a/simpleclient/src/main/java/io/prometheus/client/TimeWindowQuantiles.java
+++ b/simpleclient/src/main/java/io/prometheus/client/TimeWindowQuantiles.java
@@ -15,8 +15,8 @@ class TimeWindowQuantiles {
 
   private final Quantile[] quantiles;
   private final ConcurrentLinkedQueue<CKMSQuantiles> buckets;
-  private final AtomicLong lastRotateTimestampMillis;
-  private final long durationBetweenRotatesMillis;
+  private final AtomicLong lastRotateTimestampNanos;
+  private final long durationBetweenRotatesNanos;
 
   public TimeWindowQuantiles(Quantile[] quantiles, long maxAgeSeconds, int ageBuckets) {
     this.quantiles = quantiles;
@@ -24,8 +24,8 @@ class TimeWindowQuantiles {
     for (int i = 0; i < ageBuckets; i++) {
       this.buckets.add(new CKMSQuantiles(quantiles));
     }
-    this.lastRotateTimestampMillis = new AtomicLong(System.currentTimeMillis());
-    this.durationBetweenRotatesMillis = TimeUnit.SECONDS.toMillis(maxAgeSeconds) / ageBuckets;
+    this.lastRotateTimestampNanos = new AtomicLong(System.nanoTime());
+    this.durationBetweenRotatesNanos = TimeUnit.SECONDS.toNanos(maxAgeSeconds) / ageBuckets;
   }
 
   public double get(double q) {
@@ -34,7 +34,7 @@ class TimeWindowQuantiles {
   }
 
   public void insert(double value) {
-    // On concurrent `insert` and `rotate`, it might be acceptable to lose the measurement in the newest `bucket`.
+    // On concurrent `insert` and `rotate`, it should be acceptable to lose the measurement in the newest `bucket`.
     rotate();
 
     for (CKMSQuantiles bucket : buckets) {
@@ -52,16 +52,16 @@ class TimeWindowQuantiles {
     // On concurrent `rotate` and `rotate`:
     //  - `currentTime` is cached to reduce thread contention.
     //  - `lastRotate` is used to ensure the correct number of rotations.
-    long currentTime = System.currentTimeMillis();
-    long lastRotate = lastRotateTimestampMillis.get();
-    while (currentTime - lastRotate > durationBetweenRotatesMillis) {
-      if (lastRotateTimestampMillis.compareAndSet(lastRotate, lastRotate + durationBetweenRotatesMillis)) {
+    long currentTime = System.nanoTime();
+    long lastRotate = lastRotateTimestampNanos.get();
+    while (currentTime - lastRotate > durationBetweenRotatesNanos) {
+      if (lastRotateTimestampNanos.compareAndSet(lastRotate, lastRotate + durationBetweenRotatesNanos)) {
         CKMSQuantiles bucket = new CKMSQuantiles(quantiles);
         // rotate buckets (not atomic)
         buckets.add(bucket);
         buckets.remove();
       }
-      lastRotate = lastRotateTimestampMillis.get();
+      lastRotate = lastRotateTimestampNanos.get();
     }
   }