Idea:
> We could implement `TimeWindowQuantiles#rotate` with atomics (`compare-and-swap`, `retry-loop`), as `buckets` are only rotated occasionally (every 2 minutes by default).

Pros:
* avoids suspending threads for highly concurrent workloads (`get`, `insert`)
* reduces constant overhead (`get`, `insert`)

Cons:
* introduces busy waiting and retry-loop (`rotate`)
* ~reduces locality (`CKMSQuantiles[]` vs `ConcurrentLinkedQueue<CKMSQuantiles>`)~

Neutral:
* changes thread contention characteristics (`synchronized` vs `CAS & retry`)

Please check the code comments for considerations about correctness.

Relates to #480 , #481