@rrakos-evo 2019-06-05
It looks like this doesn't make a significant difference. 🙁

Before - 4e0e752 - `maxAgeSeconds=600, ageBuckets=5`:
```
Benchmark                                                           Mode  Cnt       Score         Error  Units
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4     219.757 ±      61.084  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4      81.192 ±      42.157  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4      71.233 ±      10.217  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  544183.558 ±  673721.964  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  604961.225 ± 1102601.976  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  611718.456 ±  961269.567  ns/op
```

After - 37d8303f - `maxAgeSeconds=600, ageBuckets=5`:
```
Benchmark                                                           Mode  Cnt       Score        Error  Units
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4     207.188 ±     55.895  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4      63.218 ±      7.410  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4      71.949 ±     22.530  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  548286.597 ± 593908.332  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  600827.505 ± 516400.134  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  573371.180 ± 492186.531  ns/op
```

Before - 4e0e752 - `maxAgeSeconds=60, ageBuckets=5`:
```
Benchmark                                                           Mode  Cnt       Score        Error  Units
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4     213.747 ±      7.583  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4      66.990 ±      6.503  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4      77.389 ±      9.192  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  259765.997 ± 358709.484  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  317509.004 ± 596762.748  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  324706.634 ± 382327.655  ns/op
```

After - 37d8303f - `maxAgeSeconds=60, ageBuckets=5`:
```
Benchmark                                                           Mode  Cnt       Score        Error  Units
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4     232.171 ±    141.777  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4      68.657 ±     29.625  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4      67.339 ±     22.754  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  337335.156 ± 125627.195  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  326169.746 ± 121918.667  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  244847.130 ± 206621.951  ns/op
```

Before - 4e0e752 - `maxAgeSeconds=6, ageBuckets=5`:
```
Benchmark                                                           Mode  Cnt       Score       Error  Units
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4     246.210 ±    17.015  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4      67.433 ±    10.148  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4      70.137 ±    10.208  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  106283.964 ± 59817.072  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  117276.079 ± 15151.896  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  112516.505 ± 48298.023  ns/op
```

After - 37d8303f - `maxAgeSeconds=6, ageBuckets=5`:
```
Benchmark                                                           Mode  Cnt       Score       Error  Units
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4     222.550 ±    29.739  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4      67.279 ±     8.880  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4      69.566 ±    12.827  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  108556.901 ± 58692.370  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  116191.326 ± 22354.193  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  115664.165 ± 22821.461  ns/op
```

Before - 4e0e752 - `maxAgeSeconds=1, ageBuckets=5`:
```
Benchmark                                                           Mode  Cnt      Score       Error  Units
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4    226.792 ±    51.313  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4     67.871 ±    10.559  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4     68.260 ±     6.567  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  45594.702 ± 12321.879  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  48994.075 ±  3317.460  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  47073.086 ±  5388.792  ns/op
```

After - 37d8303f - `maxAgeSeconds=1, ageBuckets=5`:
```
Benchmark                                                           Mode  Cnt      Score      Error  Units
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4    204.281 ±   33.846  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4     61.800 ±    5.339  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4     68.755 ±    6.382  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  48351.186 ± 5438.897  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  48931.374 ± 9847.791  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  47329.281 ± 6492.838  ns/op
```

It makes sense to me that more frequent rotations (shorter time window) is faster, as there are less measurements and samples (for `CKMSQuantiles`).